Refactor the initial workflow
Remove the page with the join link when creating a new game session -
put a link into the configure page to make the workflow simpler.
Refactor the domain to support the changes. Implement handling various
states on various pages when loading the page and navigating through the
browser's back/forward buttons.