Make all flows pure and move side effects to commands
Introduce a monad - Flow - which is a combination of the reader and
writer monads. The reader part is for reading the usual environment and
the writer part is for adding commands which are to be executed by
Bolero when the flow ends. Move all side effects, like working with
local storage or the event mediator, to commands, thus making all flows
pure - they just return the effects along with the new model.

Add a new message type - Custom - to easily implement flow continuations
when there's async stuff going on in the middle (which requires a
command). It's custom because the Flow type is defined above the message
type, and I didn't want to make them mutually recursive, so it accepts
anything (in the form of obj). The update function only upderstands
custom flow continuations.